+++ src/App.jsx
@@ -1,6 +1,6 @@
 import './App.css';
-import TodoList from './TodoList.jsx';
-import TodoForm from './TodoForm.jsx';
+import TodoList from './features/TodoList/TodoList.jsx';
+import TodoForm from './features/TodoForm.jsx';
 import { useState } from 'react';
 
 function App() {
